Move PopulationSliders type definition
